+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 * Initial import
 * Fix problem with default status
+* Raise `SchemaValidationFailed` exception whether body is broken